Don't allow to drop on the folder itself. <BTS:985>
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@2384 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qm/src/ui/folderwindow.cpp
@@ -674,7 +674,7 @@ void qm::FolderWindowImpl::drop(const DropTargetDropEvent& event)
 			
 			Folder* pFolder = FolderDataObject::get(pDataObject, pDocument_).second;
 			if (pFolder && pFolder->getAccount() == pAccount &&
-				(!pTarget || !pFolder->isAncestorOf(pTarget))) {
+				(!pTarget || (pTarget != pFolder && !pFolder->isAncestorOf(pTarget)))) {
 				if (pTarget && pTarget->isFlag(Folder::FLAG_TRASHBOX)) {
 					if (!FolderDeleteAction::deleteFolder(pFolderModel_, pFolder))
 						messageBox(Application::getApplication().getResourceHandle(),
@@ -1034,7 +1034,7 @@ void qm::FolderWindowImpl::processDragEvent(const DropTargetDragEvent& event)
 			
 			Folder* pFolder = FolderDataObject::get(pDataObject, pDocument_).second;
 			if (pFolder && pFolder->getAccount() == pAccount &&
-				(!pTarget || !pFolder->isAncestorOf(pTarget)))
+				(!pTarget || (pTarget != pFolder && !pFolder->isAncestorOf(pTarget))))
 				dwEffect = DROPEFFECT_MOVE;
 		}
